requirements.txt
# Market data
yfinance>=0.2.40

# Data handling
pandas>=2.0
numpy>=1.24
pyyaml>=6.0

# HTTP + HTML parsing for SEC EDGAR
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9

# Visualization
matplotlib>=3.7
seaborn>=0.13

# CLI
click>=8.1

# Optional NLP for sentiment (loaded lazily)
# transformers>=4.40
# torch>=2.0
